drivers: modem: sim7080: own file for pdp handling
PDP context handling takes place in separate source file for better
readability.
