+++ drivers/modem/simcom/sim7080/CMakeLists.txt
@@ -8,6 +8,7 @@ zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/ip)
 zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/lib/sockets)
 zephyr_library_sources(
     sim7080.c
+    sim7080_pdp.c
     sim7080_dns.c
     sim7080_sock.c
     sim7080_sms.c
